+++ source/TR-181/sbapi/wifi_monitor.c
@@ -1974,7 +1974,7 @@ static void upload_client_debug_stats_sta_fa_info(INT apIndex, sta_data_t *sta)
     
     memset (buf, 0, CLIENT_STATS_MAX_LEN_BUF);
     if (sta != NULL) {
-        fp = (FILE *)v_secure_popen("r", "dmesg | grep FA_INFO_%s | tail -1", to_sta_key(sta->sta_mac, sta_key));
+        fp = (FILE *)(long int)v_secure_popen("r", "dmesg | grep FA_INFO_%s | tail -1", to_sta_key(sta->sta_mac, sta_key));
         if (fp) {
             fgets(buf, CLIENT_STATS_MAX_LEN_BUF, fp);
             v_secure_pclose(fp);
@@ -2049,7 +2049,7 @@ static void upload_client_debug_stats_sta_fa_lmac_data_stats(INT apIndex, sta_da
     memset (buf, 0, CLIENT_STATS_MAX_LEN_BUF);
 
     if (sta != NULL) {
-        fp = (FILE *)v_secure_popen("r", "dmesg | grep FA_LMAC_DATA_STATS_%s | tail -1", to_sta_key(sta->sta_mac, sta_key));
+        fp = (FILE *)(long int)v_secure_popen("r", "dmesg | grep FA_LMAC_DATA_STATS_%s | tail -1", to_sta_key(sta->sta_mac, sta_key));
         if (fp) {
             fgets(buf, CLIENT_STATS_MAX_LEN_BUF, fp);
             v_secure_pclose(fp);
@@ -2114,7 +2114,7 @@ static void upload_client_debug_stats_sta_fa_lmac_mgmt_stats(INT apIndex, sta_da
 
     memset (buf, 0, CLIENT_STATS_MAX_LEN_BUF);
     if(sta != NULL) {
-        fp = (FILE *)v_secure_popen("r", "dmesg | grep FA_LMAC_MGMT_STATS_%s | tail -1", to_sta_key(sta->sta_mac, sta_key));
+        fp = (FILE *)(long int)v_secure_popen("r", "dmesg | grep FA_LMAC_MGMT_STATS_%s | tail -1", to_sta_key(sta->sta_mac, sta_key));
         if (fp) {
             fgets(buf, CLIENT_STATS_MAX_LEN_BUF, fp);
             v_secure_pclose(fp);
@@ -2180,7 +2180,7 @@ static void upload_client_debug_stats_sta_vap_activity_stats(INT apIndex)
     if (0 == apIndex) {
 
         memset (buf, 0, CLIENT_STATS_MAX_LEN_BUF);
-        fp = (FILE *)v_secure_popen("r", "dmesg | grep VAP_ACTIVITY_ath0 | tail -1");
+        fp = (FILE *)(long int)v_secure_popen("r", "dmesg | grep VAP_ACTIVITY_ath0 | tail -1");
         if (fp)
         {
             fgets(buf, CLIENT_STATS_MAX_LEN_BUF, fp);
@@ -2227,7 +2227,7 @@ static void upload_client_debug_stats_sta_vap_activity_stats(INT apIndex)
     if (1 == apIndex) {
         memset (buf, 0, CLIENT_STATS_MAX_LEN_BUF);
 
-        fp = (FILE *)v_secure_popen("r", "dmesg | grep VAP_ACTIVITY_ath1 | tail -1");
+        fp = (FILE *)(long int)v_secure_popen("r", "dmesg | grep VAP_ACTIVITY_ath1 | tail -1");
         if (fp)
         {
             fgets(buf, CLIENT_STATS_MAX_LEN_BUF, fp);
